+++ molten/validation/field.py
@@ -15,6 +15,7 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+import re
 from typing import (
     Any, Callable, Dict, Generic, List, Optional, Sequence, Type, TypeVar, Union, no_type_check
 )
@@ -242,11 +243,15 @@ class StringValidator:
             field: Field[_T],
             value: str,
             choices: Optional[Sequence[str]] = None,
+            pattern: Optional[str] = None,
             min_length: Optional[int] = None,
             max_length: Optional[int] = None,
     ) -> str:
         if choices is not None and value not in choices:
             raise FieldValidationError(f"must be one of: {', '.join(repr(choice) for choice in choices)}")
+
+        if pattern is not None and not re.match(pattern, value):
+            raise FieldValidationError(f"must match pattern {pattern!r}")
 
         if min_length is not None and len(value) < min_length:
             raise FieldValidationError(f"length must be >= {min_length}")